Update detection with hw3
